Refactor rest plugin routes
The rest plugin routes were added in the db update with curl requests. These requests are moved
to an own class. In addition the new v2 routes are added and the old db update steps are refactored
to the new classes. The refactoring does not affect the update step behavior, therefore its safe to use it.